refactor how capture options are calculated
