+++ toolkit/featureforms/src/main/java/com/arcgismaps/toolkit/featureforms/internal/components/attachment/AttachmentFormElement.kt
@@ -16,34 +16,18 @@
 
 package com.arcgismaps.toolkit.featureforms.internal.components.attachment
 
-import android.net.Uri
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Add
-import androidx.compose.material.icons.rounded.Folder
-import androidx.compose.material.icons.rounded.Photo
-import androidx.compose.material.icons.rounded.PhotoCamera
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -52,16 +36,12 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.listSaver
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.composed
-import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.geometry.Offset
@@ -69,29 +49,19 @@ import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
 import com.arcgismaps.mapping.featureforms.FormAttachmentType
-import com.arcgismaps.toolkit.featureforms.R
-import com.arcgismaps.toolkit.featureforms.internal.utils.AttachmentsFileProvider
-import com.arcgismaps.toolkit.featureforms.internal.utils.DialogType
-import com.arcgismaps.toolkit.featureforms.internal.utils.LocalDialogRequester
+import com.arcgismaps.toolkit.featureforms.internal.components.base.ValidationErrorState
 import com.arcgismaps.toolkit.featureforms.theme.AttachmentsElementColors
 import com.arcgismaps.toolkit.featureforms.theme.AttachmentsElementTypography
 import com.arcgismaps.toolkit.featureforms.theme.LocalColorScheme
 import com.arcgismaps.toolkit.featureforms.theme.LocalTypography
-import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
-import java.time.Instant
-import androidx.core.net.toUri
-import com.arcgismaps.toolkit.featureforms.internal.components.base.ValidationErrorState
 
 @Composable
 internal fun AttachmentFormElement(
@@ -100,13 +70,11 @@ internal fun AttachmentFormElement(
 ) {
     val context = LocalContext.current
     val editable by state.isEditable.collectAsState()
-    val inputType: ImageAttachmentsFormInput = state.inputType
     AttachmentFormElement(
         label = state.label,
         description = state.description,
         isEditable = editable,
-        captureOptions = CaptureOptions.Any,
-        inputType = inputType,
+        inputType = state.inputType,
         hasCameraPermission = state.hasCameraPermissions(context),
         allowUserRename = state.allowUserRename,
         displayFilename = state.displayFilename,
@@ -127,8 +95,7 @@ internal fun AttachmentFormElement(
     label: String,
     description: String,
     isEditable: Boolean,
-    captureOptions: CaptureOptions,
-    inputType: ImageAttachmentsFormInput,
+    inputType: AttachmentsFormInput,
     hasCameraPermission: Boolean,
     allowUserRename: Boolean,
     displayFilename: Boolean,
@@ -172,7 +139,7 @@ internal fun AttachmentFormElement(
                     AddAttachment(
                         onFocused = onFocused,
                         stateId = stateId,
-                        captureOptions = captureOptions,
+                        //captureOptions = captureOptions,
                         inputType = inputType,
                         hasCameraPermission = hasCameraPermission
                     )
@@ -284,281 +251,6 @@ private fun Header(
     }
 }
 
-@Composable
-private fun AddAttachment(
-    onFocused: () -> Unit,
-    stateId: Int,
-    captureOptions: CaptureOptions,
-    inputType: ImageAttachmentsFormInput,
-    hasCameraPermission: Boolean,
-) {
-    var showMenu by remember { mutableStateOf(false) }
-    val dialogRequester = LocalDialogRequester.current
-    val scope = rememberCoroutineScope()
-    val pickerStyle = remember { MutableSharedFlow<PickerStyle>() }
-    val inputMethod = inputType.inputMethod
-    val supportsCapture = remember(inputMethod) {
-        inputMethod == ImageAttachmentsFormInput.InputMethod.Capture ||
-            inputMethod == ImageAttachmentsFormInput.InputMethod.Any
-    }
-
-    val supportsUpload = remember(inputMethod) {
-        inputMethod == ImageAttachmentsFormInput.InputMethod.Upload ||
-            inputMethod == ImageAttachmentsFormInput.InputMethod.Any
-    }
-
-    Box {
-        IconButton(
-            onClick = {
-                onFocused()
-                showMenu = true
-            },
-        ) {
-            Icon(
-                Icons.Rounded.Add,
-                contentDescription = "Add attachment",
-                modifier = Modifier.size(32.dp)
-            )
-        }
-        DropdownMenu(
-            expanded = showMenu,
-            offset = DpOffset.Zero,
-            onDismissRequest = { showMenu = false }
-        ) {
-            if (hasCameraPermission && captureOptions.hasImageCapture() && supportsCapture) {
-                DropdownMenuItem(
-                    text = { Text(text = stringResource(R.string.take_photo)) },
-                    trailingIcon = {
-                        Icon(
-                            imageVector = Icons.Rounded.PhotoCamera,
-                            contentDescription = "Take Photo",
-                            modifier = Modifier.alpha(0.4f)
-                        )
-                    },
-                    onClick = {
-                        scope.launch {
-                            pickerStyle.emit(PickerStyle.Camera)
-                            showMenu = false
-                        }
-                    }
-                )
-            }
-            if (captureOptions.hasMediaCapture() && supportsUpload) {
-                DropdownMenuItem(
-                    text = { Text(text = stringResource(R.string.add_from_gallery)) },
-                    trailingIcon = {
-                        Icon(
-                            imageVector = Icons.Rounded.Photo,
-                            contentDescription = "Add From Gallery",
-                            modifier = Modifier.alpha(0.4f)
-                        )
-                    },
-                    onClick = {
-                        scope.launch {
-                            val visualMediaType =
-                                if (captureOptions.hasImageCapture() && captureOptions.hasVideoCapture()) {
-                                    ActivityResultContracts.PickVisualMedia.ImageAndVideo
-                                } else if (captureOptions.hasImageCapture()) {
-                                    ActivityResultContracts.PickVisualMedia.ImageOnly
-                                } else {
-                                    ActivityResultContracts.PickVisualMedia.VideoOnly
-                                }
-                            pickerStyle.emit(PickerStyle.PickMedia(visualMediaType))
-                            showMenu = false
-                        }
-                    }
-                )
-            }
-            if (captureOptions.hasFileCapture() && supportsUpload) {
-                DropdownMenuItem(
-                    text = { Text(text = stringResource(R.string.add_file)) },
-                    trailingIcon = {
-                        Icon(
-                            imageVector = Icons.Rounded.Folder,
-                            contentDescription = "Add File",
-                            modifier = Modifier.alpha(0.4f)
-                        )
-                    },
-                    onClick = {
-                        scope.launch {
-                            pickerStyle.emit(PickerStyle.File(captureOptions.getAllowedMimeTypes()))
-                            showMenu = false
-                        }
-                    }
-                )
-            }
-        }
-    }
-    LaunchedEffect(Unit) {
-        pickerStyle.collect {
-            when (it) {
-                PickerStyle.Camera -> {
-                    dialogRequester.requestDialog(
-                        DialogType.ImageCaptureDialog(
-                            stateId = stateId
-                        )
-                    )
-                }
-
-                is PickerStyle.PickMedia -> {
-                    dialogRequester.requestDialog(
-                        DialogType.GalleryPickerDialog(
-                            stateId = stateId,
-                            type = it.type
-                        )
-                    )
-                }
-
-                is PickerStyle.File -> {
-                    dialogRequester.requestDialog(
-                        DialogType.FilePickerDialog(
-                            stateId = stateId,
-                            allowedTypes = it.allowedMimeTypes
-                        )
-                    )
-                }
-            }
-        }
-    }
-}
-
-/**
- * Launches the camera to capture an image. When an image is captured, the [onImageCaptured] callback
- * is invoked with the URI of the captured image. In case of a dismissal or if no image is captured,
- * the [onDismissRequest] callback is invoked.
- *
- * @param onDismissRequest A request to dismiss the camera picker.
- * @param onImageCaptured A callback to invoke when an image is captured.
- */
-@Composable
-internal fun ImageCapture(
-    onDismissRequest: () -> Unit,
-    onImageCaptured: (Uri) -> Unit
-) {
-    val context = LocalContext.current
-    var hasLaunched by rememberSaveable {
-        mutableStateOf(false)
-    }
-    val capturedImageUri = rememberSaveable(
-        saver = listSaver(
-            save = { listOf(it.toString()) },
-            restore = { it.first().toUri() }
-        )
-    ) {
-        val timeStamp = Instant.now().toEpochMilli()
-        AttachmentsFileProvider.createTempFileWithUri("IMAGE_$timeStamp", ".jpg", context)
-    }
-    val cameraLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.TakePicture(),
-        onResult = { success ->
-            if (success) {
-                onImageCaptured(capturedImageUri)
-            } else {
-                onDismissRequest()
-            }
-        }
-    )
-    LaunchedEffect(Unit) {
-        if (!hasLaunched) {
-            hasLaunched = true
-            cameraLauncher.launch(capturedImageUri)
-        }
-    }
-}
-
-/**
- * Launches the Gallery to select an image, video or both based on the [type]. When a selection is
- * made, the [onMediaSelected] callback is invoked with the URI of the selected image/video. In case
- * of a dismissal or if no media is selected, the [onDismissRequest] callback is invoked.
- *
- * @param type The type of media to select.
- * @param onDismissRequest A request to dismiss the gallery picker.
- * @param onMediaSelected A callback to invoke when a media file is selected.
- */
-@Composable
-internal fun GalleryPicker(
-    type: ActivityResultContracts.PickVisualMedia.VisualMediaType,
-    onDismissRequest: () -> Unit,
-    onMediaSelected: (Uri) -> Unit
-) {
-    var hasLaunched by rememberSaveable {
-        mutableStateOf(false)
-    }
-    val launcher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.PickVisualMedia()
-    ) {
-        if (it != null) {
-            onMediaSelected(it)
-        } else {
-            onDismissRequest()
-        }
-    }
-    Dialog(onDismissRequest = onDismissRequest) {
-        CircularProgressIndicator(
-            modifier = Modifier.size(50.dp),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            strokeWidth = 5.dp
-        )
-    }
-    LaunchedEffect(Unit) {
-        if (!hasLaunched) {
-            hasLaunched = true
-            launcher.launch(PickVisualMediaRequest(type))
-        }
-    }
-}
-
-/**
- * Launches the file picker to select a file based on the [allowedMimeTypes]. When a file is selected,
- * the [onFileSelected] callback is invoked with the URI of the selected file. In case of a dismissal
- * or if no file is selected, the [onDismissRequest] callback is invoked.
- *
- * @param allowedMimeTypes The list of allowed MIME types to select.
- * @param onDismissRequest A request to dismiss the file picker.
- * @param onFileSelected A callback to invoke when a file is selected.
- */
-@Composable
-internal fun FilePicker(
-    allowedMimeTypes: List<String>,
-    onDismissRequest: () -> Unit,
-    onFileSelected: (Uri) -> Unit
-) {
-    var hasLaunched by rememberSaveable {
-        mutableStateOf(false)
-    }
-    val launcher = rememberLauncherForActivityResult(ActivityResultContracts.OpenDocument()) {
-        if (it != null) {
-            onFileSelected(it)
-        } else {
-            onDismissRequest()
-        }
-    }
-    Dialog(onDismissRequest = onDismissRequest) {
-        CircularProgressIndicator(
-            modifier = Modifier.size(50.dp),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            strokeWidth = 5.dp
-        )
-    }
-    LaunchedEffect(Unit) {
-        if (!hasLaunched) {
-            hasLaunched = true
-            launcher.launch(allowedMimeTypes.toTypedArray())
-        }
-    }
-}
-
-/**
- * Determines the type of picker to launch.
- */
-private sealed class PickerStyle {
-    data object Camera : PickerStyle()
-    data class PickMedia(val type: ActivityResultContracts.PickVisualMedia.VisualMediaType) :
-        PickerStyle()
-
-    data class File(val allowedMimeTypes: List<String>) : PickerStyle()
-}
-
 /**
  * Creates a horizontal scrollbar for a [LazyRow] with the given [state]. [offsetY] can be used to
  * adjust the offset of the scrollbar in the Y axis. This also adds the required padding to the
@@ -651,7 +343,6 @@ private fun AttachmentFormElementPreview() {
         label = "Attachments",
         description = "Add attachments",
         isEditable = true,
-        captureOptions = CaptureOptions.Any,
         inputType = ImageAttachmentsFormInput(
             inputMethod = ImageAttachmentsFormInput.InputMethod.Capture
         ),
