+++ toolkit/featureforms/src/main/java/com/arcgismaps/toolkit/featureforms/internal/components/attachment/GenericAttachmentFormInput.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2026 Esri
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.arcgismaps.toolkit.featureforms.internal.components.attachment
+
+// Prototype
+
+/**
+ * A prototype implementation for a generic attachments form input class.
+ */
+internal class GenericAttachmentFormInput(
+    val associationType: AttachmentAssociationType,
+    val inputTypes : List<AttachmentsFormInput>
+) : AttachmentsFormInput()
+
+internal sealed class AttachmentAssociationType {
+    data object Any : AttachmentAssociationType()
+    data object Exact : AttachmentAssociationType()
+    data object ExactOrNone : AttachmentAssociationType()
+}
