+++ toolkit/featureforms/src/main/res/values/strings.xml
@@ -179,4 +179,6 @@
     <string name="value">Value</string>
     <string name="filters_not_applied">Filters have not been applied</string>
     <string name="discard_changes_confirmation">Are you sure you want to discard the changes?</string>
+    <string name="record_audio">Record Audio</string>
+    <string name="record_video">Record Video</string>
 </resources>
